Convert numeric to factor in data frame and check

These are convenience functions for license data, similar to factor_var but operate on data frames (useful for piping) and produce a check summary.

df_factor_var(df, var, levels, labels, suppress_check = TRUE, ...)

df_factor_sex(df, var = "sex", levels = 1:2, labels = c("Male",
  "Female"), ...)

df_factor_res(df, var = "res", levels = c(1, 0),
  labels = c("Resident", "Nonresident"), ...)

df_factor_R3(df, var = "R3", levels = 1:4, labels = c("Carry",
  "Renew", "Reactivate", "Recruit"), ...)

df_factor_age(df, var = "age", levels = 1:7, labels = c("0-17",
  "18-24", "25-34", "35-44", "45-54", "55-64", "65+"), ...)

Arguments

df
data frame: Input data frame

var
character: Name of numeric variable to convert

levels
numeric: Levels for input numeric vector

labels
labels: Labels to use for output factor vector

suppress_check
logical: If TRUE, does not print a coding summary

...
Other arguments passed to factor

See also

Examples

library(dplyr)
data(history)
x <- history %>%
    df_factor_R3(suppress_check = FALSE) %>%
    df_factor_res(suppress_check = FALSE) %>%
    df_factor_sex(suppress_check = FALSE)

#> # A tibble: 5 x 3
#>   new          old     n
#>   <fct>      <int> <int>
#> 1 Carry          1 10476
#> 2 Renew          2 21134
#> 3 Reactivate     3  8941
#> 4 Recruit        4 13534
#> 5 <NA>          NA 37821
#> # A tibble: 2 x 3
#>   new           old     n
#>   <fct>       <int> <int>
#> 1 Resident        1 71345
#> 2 Nonresident     0 20561
#> # A tibble: 3 x 3
#>   new      old     n
#>   <fct>  <int> <int>
#> 1 Male       1 71930
#> 2 Female     2 18474
#> 3 <NA>      NA  1502