Estimate churn by year from license history

This function requires a correctly formated history table (see history). It runs a mean of the lapse value (per year), optionally by segment (and also shifts year forward by 1 so that churn in current year reflects lapse pct from last year). It also runs a validation test: pct change per year.

est_churn(history, segment = "tot", test_threshold = 30,
  show_test_stat = FALSE, suppress_warning = FALSE, outvar = "churn")

Arguments

history
data frame: input license history table

segment
character: defaults to "tot", which indicates no segmentation. Alternatively specifiy other license history variables ("res", "sex", etc.)

test_threshold
numeric: threshold in whole number percentage points for pct change per year. A warning will be printed if the absolute value of the change for any year exceeds the threshold.

show_test_stat
logical: If TRUE, the output table will include a variable holding the test statistic for each row.

suppress_warning
logical: If TRUE, no test warning will be displayed (even if threshold is exceeded). Test statistics can still be included by setting show_test_stat = TRUE.

outvar
character: name of variable that stores metric

See also

Examples

library(dplyr)
data(history)
history <- history %>%
    label_categories() %>%
    recode_agecat() %>%
    filter(!agecat %in% c("0-17", "65+"))
est_churn(history)

#> # A tibble: 10 x 3
#>    tot    year churn
#>    <chr> <dbl> <dbl>
#>  1 All    2009 0.403
#>  2 All    2010 0.435
#>  3 All    2011 0.444
#>  4 All    2012 0.419
#>  5 All    2013 0.430
#>  6 All    2014 0.411
#>  7 All    2015 0.458
#>  8 All    2016 0.480
#>  9 All    2017 0.481
#> 10 All    2018 0.488

# apply across all segments
segs <- c("tot", "res", "sex", "agecat")
churn <- sapply(segs, function(x) est_churn(history, x), simplify = FALSE)