salic: Prepare Agency Dashboard Data

Run vignette("salic") from the R console for an introduction. See below for a summary of the core functions.

Standardized Data

see cust, lic, sale for information about expected license data formats

use data_check to ensure correct formatting for state license data

License History

use rank_sale to condense sales to 1 row per customer-year

use make_history to produce a history table that facilitates flexible querying about customer trends

see history_check to run checks for the the output of make_history()

Dashboard Metrics

use label_categories & recode_agecat to prepare the history table

use est_part to get participant counts in a metrics list

use scaleup_part to account for missing values in segments such as gender, age, etc.

use est_churn to get churn rates per year

use format_result to stack metrics into a dashboard data frame